split the Forms.jsx file into LoginForm.jsx and RegForm.jsx components
